+++ packages/react-components/src/useCombinedRefs.ts
@@ -8,7 +8,6 @@ type MutableRef<T> = MutableRefObject<T> | RefCallback<T>;
  * Combine multiple refs into one functional ref.
  *
  * @param refs - The refs to combine.
- *
  * @returns A function that will apply the value on all given refs.
  */
 export function useCombinedRefs<T>(...refs: MutableRef<T>[]): RefCallback<T> {