+++ packages/server/src/utils/test/authorization.ts
@@ -1,7 +1,7 @@
 import { randomBytes } from 'crypto';
 
 import { api } from '@appsemble/utils';
-// eslint-disable-next-line import/no-extraneous-dependencies, node/no-unpublished-import
+// eslint-disable-next-line node/no-unpublished-import
 import { request } from 'axios-test-instance';
 import { hash } from 'bcrypt';
 import { OpenAPIV3 } from 'openapi-types';