+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## \[[0.18.15](https://gitlab.com/appsemble/appsemble/-/releases/0.18.15)] - 2021-06-01
+
+### Added
+
+- Studio: Add core app message translations to the translations page.
+
+### Changed
+
+- Server: Block organization IDs that could be used to impersonate Appsemble.
+- Server: Redirect to Appsemble Studio for app-like URLs.
+
 ## \[[0.18.14](https://gitlab.com/appsemble/appsemble/-/releases/0.18.14)] - 2021-05-26
 
 ### Fixed